Some folders organization changes
